embed_kernel: only import zmq when needed
